fix(lib): move plugin source of fromData into corrrect file
